pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ats-screen"
version = "1.0.0"
description = "AI-powered resume screening — multi-agent CLI built on the Claude Agent SDK."
readme = "README.md"
requires-python = ">=3.10,<3.15"
license = "MIT"
authors = [{ name = "Ali Razmdideh" }]
keywords = ["ats", "resume", "screening", "agents", "claude", "anthropic", "llm"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "claude-agent-sdk>=0.1.69",
    "anyio>=4.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pypdf>=4.0",
    "python-docx>=1.1",
    "rapidfuzz>=3.6",
    "rich>=13.7",
    "typer>=0.12",
    "httpx>=0.27",
    "tenacity>=9.0",
    "python-json-logger>=3.0",
]

[project.urls]
Homepage = "https://github.com/aliRazmdideh/ats"
Repository = "https://github.com/aliRazmdideh/ats"
Issues = "https://github.com/aliRazmdideh/ats/issues"

[project.optional-dependencies]
dev = [
    "black==24.10.0",
    "isort==5.13.2",
    "flake8==7.1.1",
    "mypy==1.13.0",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "build>=1.2",
    "fpdf2>=2.7",
]

[project.scripts]
ats = "ats.cli:main"

[tool.setuptools.packages.find]
include = ["ats*"]
exclude = ["tests*"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
disallow_untyped_decorators = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
